requirements.txt
captchatools==1.4.1
curl-cffi==0.7.4
pymongo==4.10.1
python-dotenv==1.0.1
colorama==0.4.6
rich==13.9.4
